+++ gazebo/physics/bullet/BulletSliderJoint.cc
@@ -19,6 +19,7 @@
 #include "gazebo/common/Exception.hh"
 
 #include "gazebo/physics/Model.hh"
+#include "gazebo/physics/World.hh"
 #include "gazebo/physics/bullet/bullet_inc.h"
 #include "gazebo/physics/bullet/BulletLink.hh"
 #include "gazebo/physics/bullet/BulletPhysics.hh"
@@ -390,8 +391,16 @@ bool BulletSliderJoint::SetParam(const std::string &_key,
       {
         this->bulletSlider->setPoweredLinMotor(true);
         this->bulletSlider->setTargetLinMotorVelocity(0.0);
+        // there is an upstream bug in bullet 2.82 and earlier
+        // the maxLinMotorForce parameter is inadvertently divided
+        // by timestep when attempting to compute an impulse in
+        // the bullet solver.
+        // https://github.com/bulletphysics/bullet3/pull/328
+        // As a workaround, multiply the desired friction
+        // parameter by dt^2 when setting
+        double dt = this->world->GetPhysicsEngine()->GetMaxStepSize();
         this->bulletSlider->setMaxLinMotorForce(
-          boost::any_cast<double>(_value));
+          dt*dt * boost::any_cast<double>(_value));
       }
       else
       {
@@ -427,7 +436,15 @@ double BulletSliderJoint::GetParam(const std::string &_key, unsigned int _index)
   {
     if (this->bulletSlider)
     {
-      return this->bulletSlider->getMaxLinMotorForce();
+      // there is an upstream bug in bullet 2.82 and earlier
+      // the maxLinMotorForce parameter is inadvertently divided
+      // by timestep when attempting to compute an impulse in
+      // the bullet solver.
+      // https://github.com/bulletphysics/bullet3/pull/328
+      // As a workaround, divide the desired friction
+      // parameter by dt^2 when getting
+      double dt = this->world->GetPhysicsEngine()->GetMaxStepSize();
+      return this->bulletSlider->getMaxLinMotorForce() / (dt*dt);
     }
     else
     {